requirements.txt
langchain
langchain-core
langchain-google-genai
python-dotenv
streamlit
pydantic